move parse to common
