+++ core/src/commands/follow_chain.rs
@@ -28,7 +28,9 @@ use sp_runtime::{
 use substrate_rpc_client::{ws_client, ChainApi, FinalizedHeaders, Subscription, WsClient};
 
 use crate::{
-    build_executor, full_extensions, parse, rpc_err_handler,
+    build_executor,
+    common::parse,
+    full_extensions, rpc_err_handler,
     state::{LiveState, RuntimeChecks, State},
     state_machine_call_with_proof, SharedParams, LOG_TARGET,
 };
