+++ core/src/commands/offchain_worker.rs
@@ -23,7 +23,9 @@ use sp_runtime::traits::{Block as BlockT, NumberFor};
 use substrate_rpc_client::{ws_client, ChainApi};
 
 use crate::{
-    build_executor, full_extensions, parse, rpc_err_handler,
+    build_executor,
+    common::parse,
+    full_extensions, rpc_err_handler,
     state::{LiveState, RuntimeChecks, State},
     state_machine_call, SharedParams, LOG_TARGET,
 };
