+++ core/src/common/mod.rs
@@ -1,2 +1,3 @@
 pub mod misc_logging;
+pub mod parse;
 pub mod shared_parameters;
